Revert "Tup 19 tarea test"
