+++ .github/workflows/FrontCi.yml
@@ -1,9 +1,12 @@
-name: FrontCI Pipeline
+name: ForntCI Pipeline
 
 on:
   push:
     branches:
-      - '**' # Se ejecuta en todas las ramas
+      - main # Es pora que se ejecute en main
+  pull_request:
+    branches:
+      - '**' # Es pora que se ejecute en todas las ramas
 
 jobs:
   lint:
@@ -26,31 +29,3 @@ jobs:
         run: npm run lint
         working-directory: ./FrontAdmin
 
-  prettier:
-    name: Formatear código con Prettier
-    runs-on: ubuntu-latest
-    if: github.ref == 'refs/heads/main' # Solo se ejecuta en la rama main
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v4
-
-      - name: Setear Node.js
-        uses: actions/setup-node@v4
-        with:
-          node-version: 20
-
-      - name: Instalar dependencias
-        run: npm install
-        working-directory: ./FrontAdmin
-
-      - name: Correr Prettier
-        run: npm run format
-        working-directory: ./FrontAdmin
-
-      - name: Commit changes
-        run: |
-          git config --local user.name "TobiasMaciel"
-          git config --local user.email "tobiasmaciel03@gmail.com"
-          git add .
-          git commit -m "chore: Se formateo el codigo con Prettier" || echo "No hay cambios para comitear"
-          git push
