+++ .gitignore
@@ -48,7 +48,5 @@ __pycache__/
 
 # API Docs
 
-*schema.yml
-
 # sql files
 *.sql
