+++ FrontAdmin/.eslintrc.cjs
@@ -5,20 +5,17 @@ module.exports = {
     "eslint:recommended",
     "plugin:@typescript-eslint/recommended",
     "plugin:react-hooks/recommended",
-    "plugin:prettier/recommended"
   ],
   ignorePatterns: ["dist", ".eslintrc.cjs"],
   parser: "@typescript-eslint/parser",
-  plugins: ["react-refresh", "prettier"],
+  plugins: ["react-refresh"],
   rules: {
-    "prettier/prettier": "warn", // Se muestra los errores de Prettier como warnings
     "react-refresh/only-export-components": [
       "warn",
       { allowConstantExport: true },
     ],
-    "@typescript-eslint/prefer-as-const": "warn",
-    "@typescript-eslint/no-unused-vars": "warn",
-    "no-extra-semi": "warn",
-    "no-mixed-spaces-and-tabs": "warn",
+    "@typescript-eslint/prefer-as-const": "warn", // Warning de la regla que requiere const en lugar de type assertion
+    "@typescript-eslint/no-unused-vars": "warn", // Warning de la regla que indica variables no usadas
+    "no-extra-semi": "warn", // Warning de la regla que prohíbe los puntos y comas innecesarios
   },
 };
