--- FrontAdmin/.prettierrc
@@ -1,9 +1,0 @@
-{
-    "semi": true,
-    "singleQuote": true,
-    "trailingComma": "es5",
-    "printWidth": 80,
-    "tabWidth": 2,
-    "useTabs": false
-  }
-
