+++ FrontAdmin/src/App.tsx
@@ -3,7 +3,7 @@ import { useEffect } from "react";
 import { BrowserRouter, Navigate, Route, Routes } from "react-router-dom";
 import { useAuth } from "./Context";
 import Auth from "./layouts/Auth";
-import Admin from "./layouts/Admin";
+import NavBar from "./components/NavBar/NavBar";
 import theme from "./theme/theme";
 
 
@@ -18,11 +18,11 @@ function App() {
   return (
     <ChakraProvider theme={theme}>
          <BrowserRouter>
-         <Routes>
+            <Routes>
               {!isAuthenticated && <Route path="/auth/*" element={<Auth />} />}
-              {isAuthenticated && <Route path="/admin/*" element={<Admin/>}/>} 
+              {isAuthenticated && <Route path="/admin/*" element={<NavBar/>} />} 
               <Route path="/*" element={<Navigate replace to={isAuthenticated ? "/admin" : "/auth"}/>} />
-          </Routes>
+            </Routes>
       </BrowserRouter> 
     </ChakraProvider>
   )
