+++ FrontAdmin/src/components/NavBar/LinksItems.tsx
@@ -3,10 +3,10 @@ import iconChat from '../icons/burbuja-de-chat 1.png';
 import iconConfig from '../icons/configuracion 1.png';
 import iconEstadisticas from '../icons/grafico-de-barras 1.png';
 
-
 export const LINK_ITEMS = [
   { icon: iconEstadisticas, url: "estadisticas" },
   { icon: iconAlumno, url: "error404" },
   { icon: iconChat, url: "error404" },
-  { icon: iconConfig, url: "configuración" },
+  { icon: iconConfig, url: "error404" },
+  
 ];
