--- FrontAdmin/src/components/Pages/Configuracion/CompromisoP.tsx
@@ -1,8 +1,0 @@
-import { Box } from "@chakra-ui/react";
-import ModalCompromisoPago from "./ModalVerDocumento";
-
-export default function CompromisoP(){
-    return(
-       <Box>-</Box>
-    );
-}
