+++ FrontAdmin/src/components/Pages/Estadisticas/SubPages/Calendario.tsx
@@ -72,6 +72,6 @@ function Calendario() {
       </Box>
   </Flex>
     )
-}
+};
 
 export default Calendario;
