+++ FrontAdmin/src/components/Pages/Login/LoginPage.tsx
@@ -27,12 +27,11 @@ function LoginPage() {
 
   const titleColor = useColorModeValue("black", "white");
   const textColor = useColorModeValue("black", "white");
-  const bgColor = useColorModeValue("#EEEEF0", "gray.800");
 
   const handleSubmit = async (e: { preventDefault: () => void; }) => {
     e.preventDefault();
     try {
-      //await FetchLogin(username, password);
+      await FetchLogin(username, password);
       onLogin();
 
     } catch (error) {
@@ -70,8 +69,8 @@ function LoginPage() {
               </Flex>
               <Heading 
                 color={titleColor} 
-                fontSize={{base:'28px',  md: "30px", lg: "35px" }}
-                m={{base:'10px',  md: "40px"}}
+                fontSize='35px'
+                m='40px'
                 fontWeight='500' outline='none'>
                 Iniciar sesión
               </Heading>
@@ -143,7 +142,7 @@ function LoginPage() {
                 </Button>
                 <Flex alignItems='center' justifyContent='center'> 
                   <Link 
-                    color={textColor} 
+                    color='textColor' 
                     href='#' 
                     fontSize='15px'
                     _hover={{ border:'none'}}
