+++ FrontAdmin/src/theme/components/button.ts
@@ -1,11 +1,7 @@
-import { background } from "@chakra-ui/react";
-import { color } from "chart.js/helpers";
-import { ImTextColor } from "react-icons/im";
-
 export const buttonStyles = {
   components: {
     Button: {
-      /*variants: {
+      variants: {
         "no-hover": {
           _hover: {
             boxShadow: "none",
@@ -28,18 +24,8 @@ export const buttonStyles = {
             boxShadow: "none",
           },
         },
-      },*/
+      },
       baseStyle: {
-        //add
-        bg: '#022855',
-        background:  "#022855",
-        fontWeight: '500',
-        color: 'white',
-        _hover:{
-          bg:  "#0f183f",
-          background:  "#0f183f",
-        },
-        //
         borderRadius: "15px",
         _focus: {
           boxShadow: "none",
