+++ FrontAdmin/src/theme/styles.ts
@@ -1,19 +1,19 @@
 import { mode, StyleFunctionProps } from '@chakra-ui/theme-tools';
 
 export const globalStyles = {
-  colors: {
-    transparent: 'transparent',
-    black: '#000',
-    white: '#fff',
+	colors: {
+		transparent: "transparent",
+    black: "#000",
+    white: "#fff",
     gray: {
-      50: '#f7fafc',
-      900: '#171923',
+      50: "#f7fafc",
+      900: "#171923",
     },
 	},
 	styles: {
-		global: (props: StyleFunctionProps) => ({
+		global: (props: StyleFunctionProps | Record<string, any>) => ({
 			body: {
-				bg: mode('eeeef0', 'gray.800')(props), //color de fondo de la pagina completa 
+				bg: mode('gray.50', 'gray.800')(props), //color de fondo de la pagina completa 
 				fontFamily: "'Roboto', sans-serif"
 			},
 			html: {
