+++ FrontAdmin/vite.config.ts
@@ -5,9 +5,4 @@ import react from '@vitejs/plugin-react-swc'
 export default defineConfig({
   plugins: [react()],
   base: '/',
-  server : {
-    hmr: {
-      overlay: false,
-    },
-  },
 })
