--- backend/core/modelos/model_mensajeria.py
@@ -1,12 +1,0 @@
-# core/models/mansajeria.py
-
-# django imports
-
-from django.db import models
-
-class Persona(models.Model):
-    pass
-class Mensaje(models.Model):
-    pass
-class DestinatarioMensaje(models.Model):
-    pass
