+++ backend/server/urls.py
@@ -27,10 +27,8 @@ urlpatterns: list[URLResolver] = [
 ]
 if settings.DEBUG:
     from debug_toolbar.toolbar import debug_toolbar_urls
-    from django.conf.urls.static import static
     from django.contrib import admin
 
-    urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
     urlpatterns += [
         path(
             "",
