--- docker-compose.yml
@@ -1,16 +1,0 @@
-version: "3.8"
-
-services:
-  web:
-    build: .
-    command: gunicorn server.wsgi:application --bind 0.0.0.0:8000
-    volumes:
-      - ./backend:/app
-      - media_volume:/app/media
-    ports:
-      - "8000:8000"
-    env_file:
-      - .env
-
-volumes:
-  media_volume:
